backend/recommendations/requirements.txt
flask==2.2.5
flask-cors==3.0.10
scikit-surprise==1.1.3
pandas==1.5.3
numpy==1.24.4
